Finished the project
I didn't bother doing the extra JS as this
will be covered in later lessons and I am
focusing on doing the HTML and CSS.

I realised that when pressing tab it goes to
the next line not the next box. This is because
I coded it in columns not in rows - would fix
this going forward.

Also the password fields don't match each other
as this was in JS.

When submitting information it doesn't get sent
anywhere but backend will be covered later.

Finally - there is a slight issue when resizing
to the smallest it can be but tbh this would
never happen in real life and mobile is a
different matter.